diffAlowance test
